Rename clinical trials module name
