+++ src/trialsynth/ctgov/config.py
