+++ src/trialsynth/ctgov/ground.py
